refactor(mobile): decompose the chat screen; fix resume height desync
The chat screen had grown to one 822-line file mixing message rendering,
the composer, keyboard/scroll choreography, one-time history bootstrap, and
tool-navigation logic. Split it into focused pieces with no behaviour or
styling change (comments preserved verbatim):

- shared/chat/tool-ref.ts (+12 unit tests) — the toolRef/NOTE_TOOLS/CARD_TOOLS
  mapping, now framework-free and testable in root vitest; mobile consumes it
  via @cairn/shared/chat/tool-ref. Not unified with the divergent desktop copies.
- mobile/src/chat/history.ts — loadInitialChat() reads on-device history ONCE
  (was read twice per mount) and seeds both the UI bubbles and the agent
  conversation.
- mobile/src/chat/useChatScroll.ts — all keyboard/scroll choreography (resume
  repaint, follow-end, composer-height -> transcript padding, attach-menu
  counter-transform).
- mobile/src/components/chat/{MessageBubble,ReasoningBlock,ToolTrail,Composer}.tsx
  — presentational, each with its own byte-identical style slice.

Screen is now ~370 lines: state + the send stream loop + composition.

Also fixes a pre-existing resume bug (predates this refactor): the ghost-band
repaint set key={resumeKey} on the KeyboardChatScrollView itself, remounting
it on background->active and resetting the library's internal layout/size
shared values to {0,0}. onContentSizeChange doesn't reliably re-fire for
identical content, so the keyboard engine read height 0 and couldn't lift
content above the keyboard until a cold relaunch. The key now sits on an inner
View wrapping just the message list, so the ghost-band content still remounts
while the scroll view keeps its measured size.
