+++ mobile/changelogs/v0.1.3.md
@@ -22,6 +22,7 @@ Polish for the workspace switcher and Knowledge Graph controls.
 - **Your AI endpoint no longer resets when you restart the app**: The custom OpenAI-compatible endpoint URL, model, and provider choice are now stored device-wide instead of inside each workspace's local database. Previously they lived in the per-workspace DB, which gets rebuilt from sync — so upgrading the app or switching workspaces silently reverted your endpoint to the default (your API key survived, which was the tell). Existing settings are migrated automatically on first launch.
 - **Chat opens at the latest message and stays the right size**: The Chat transcript now jumps to the newest message when you open the tab (and when the app resumes), instead of sometimes opening scrolled to the top and only snapping down after you sent a message. The message list also gets a stable height so it tracks content correctly rather than mis-sizing under the keyboard/composer.
 - **No more ghost bands after reopening Chat**: Returning to the Chat screen after the app had been backgrounded could briefly show faint vertical bands over the messages (an iOS snapshot artifact from the frosted composer). Chat now repaints on resume so they no longer appear.
+- **Chat scrolls correctly after the app returns from the background**: After backgrounding and reopening the app, the chat transcript could get stuck at the wrong height — you couldn't scroll the conversation up above the keyboard until you fully quit and relaunched. The resume repaint now refreshes only the messages, not the whole scroll view, so it keeps its measured size and scrolls correctly right away.
 - **Reschedule overdue tasks by dragging**: Overdue chips in the calendar's overdue tray can now be long-pressed and dragged onto a day (or into the Unscheduled tray) to reschedule or clear their due date — same as tasks in the grid.
 - **Consistent calendar header**: The Today and Month/Week controls now sit in the top navigation bar on the per-project calendar too, matching the Calendar tab — the calendar looks the same however you open it.
 - **Chat attachment icon stays put**: The image-attachment button in the chat composer no longer jumps up out of place after you tap it while the keyboard is open — it's now a plain button (opening a native photo action sheet) that rides the keyboard exactly like the send button, instead of a native menu control that anchored to the screen and drifted when its menu opened.
@@ -62,3 +63,4 @@ Polish for the workspace switcher and Knowledge Graph controls.
 - **Graph semantic edges now pool notes and cards.** `semanticEdges()` combines note and card section vectors into one corpus (shared `EmbeddingRow` shape via the `card_id AS note_id` alias), so a single all-pairs pass produces note↔note, task↔task and note↔task edges. The graph screen's existing merge (keep edges whose both endpoints are nodes) and `card`-node routing needed no changes; cards are already indexed by `catchUpIndex`.
 - Note and task semantic search now share one ranking pipeline (`rankSemantic`) parameterised by row source and id→text/title lookups, instead of two near-identical copies — so query-cache, centroid centring, hybrid dense+lexical rerank and slicing can't drift apart between notes and tasks. No behaviour change.
 - Syntax-highlighting grammars (lowlight/highlight.js `common` set) now construct lazily on the first code block render instead of at module import, so app launch and the first notes/chat render don't pay the grammar-registration cost when no code is shown.
+- Split the Chat screen — which had grown to one 800-line file mixing message rendering, the composer, keyboard/scroll choreography and tool-navigation logic — into focused pieces (a message bubble, reasoning disclosure, tool-chip trail, composer, a keyboard-scroll hook, and shared/unit-tested tool-navigation logic), and made it load its saved history only once on open instead of twice. No visible change; the chat behaves and looks exactly the same, but the code is far easier to change safely.
